refactor: move gates to separate package
